Rework CDK code
This change re-organizes and simplifies much of the CDK code. Instead of
trying to do perf and testing in one stack, they're now separated into
their own stacks.

The test stack can now accept a list of NatInstanceProvider objects as a
parameter. This will be used to deploy all AMI variants at once and test
all of them for connectivity. For now it just mimics the current
behavior to grab the latest v1.0.0 AMI.