Fixes failing attach network when already used (#43)
When using with ASG, the instance would sometimes fail to attach the ENI as it'd still be used by the other instance being terminated. This would not happen in every cases, but when it'd happen, the instance would be in a failed state where it'd wait for the interface to be attached, which would never happen as the attach network command is not retried.

When happening, manually attaching the ENI would unstuck the instance from its deadlock.