Option to use NAT instance with existing EIP (#44)
This PR gives support for static EIP by allowing instance to allocate itself an EIP passed as parameter to its public interface (eth0). This ensures network always flows through the same IP even in case of instance recreation / failure.

A few points to note in this PR:

* Once assigned, a sleep 3 is issued to ensure the instance network networking is functional as I encountered cases where a direct egress interaction following EIP change would result in failure
* I added IAM permissions on the documentation Cloudformation, without adding a way to actually pass an EIP allocation as I am fairly unfamiliar with Cloudformation
* A few user data API calls were shifted from the eni_id scope to the global scope so that those variables may be reused in other blocks (in this case eip_id)

This PR only adds support for the service to have an EIP. This change is compatible with current implementations of CDK and Cloudformation, but they will need to be updated to pass the eip_id parameter to the configuration in order to have static EIP support.